server/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
email-validator==2.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
pytest==8.1.1
ruff==0.3.4
black==24.3.0
